backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
sentence-transformers
torch
rapidfuzz
nltk
numpy
pandas
datasets
scikit-learn